Better working disassembler
Removed the PC definition, now it's just a number that is passed into the
the recur as an argument, which I think is more akin to how clojure
programs work.

Also added a check to looking up size and instruction tables so that if we
are done then we don't throw an exception and crash but rather stop
"elegantly".

Signed-off-by: pjm <p.micolet@gmail.com>