Disassembly more work
So I read more about instructions to wrap my head around it and changed the
sizes because some of it didn't make sense (implied addressing should only be
of size one). Still nowhere near sure whether or not this generates the correct result
but we should be getting there.

Also added addressing mode for each instruction to print out the proper encoding
of instruction.

Finally, added addr which, given the extra size of the instruction, fetches that data.

Signed-off-by: pjm <p.micolet@gmail.com>