New and more obfuscated addr function
Removed the increment variable because it served no real purpose.
This method does, unfortunately, require an extra run through addr function
but it's more "elegant".

Just remember that we start with a maximum size, so this means
that the first byte after the instruction is actually obtained in the second call of addr when
the size is 2.

Also, just removed some print formatting errors like not having newline
and removed the (println pc) which was there for debugging.

Signed-off-by: pjm <p.micolet@gmail.com>