Added a macro to define new instructions.
The idea was taken from the Clones NES emulator so it's not like I'm breaking
new ground. Essentially the way this works is that we define an instruction by
giving it a name, its encoding value (should be hex), data and define how the instruction
operates. This then creates a new defn with the operation and uses the instruction name as a symbol
 it then adds the symbol to the atom instr-ops.


Signed-off-by: pjm <p.micolet@gmail.com>