When doing JSR we need to keep full address.
So this is a big problem right now, going to file a todo on github.

We need a way to actually convert the PC correctly. Right now, nestest uses
a starting address of 0xC000 and most of the instructions that have to deal
with the PC fetched from memory must remove or add 0xC000 which is just a
very poor design choice, since some programs will NOT start at 0xC000.

Signed-off-by: pjm <p.micolet@gmail.com>