Huge refactoring.
This might actually break something which would suck.

Updated the def-instr macro to now accept more than one encoding. This means
that we do not need to redefine instructions but rather have a grab-data function
that will get the data according to the mode.

This means I also had to add new modes just because sometimes we only need
the raw data and not data from RAM.

Made sure this runs all the way to the same instruction I was at before
this was implemented.

Signed-off-by: pjm <p.micolet@gmail.com>